basically passed kvraft 3A tests

+++ src/kvraft/common.go
@@ -14,7 +14,7 @@ type PutAppendArgs struct {
 	Value     string
 	Op        string // "Put" or "Append"
 	ClientId  int64
-	RequestId int
+	RequestId int64
 }
 
 type PutAppendReply struct {
@@ -25,7 +25,7 @@ type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
 	ClientId  int64
-	RequestId int
+	RequestId int64
 }
 
 type GetReply struct {
